feat(plugin): add plugin to emit trace metadata
As a preparation to align the ftrace trace and the LTTng US trace, we
need a way to gather information about the clock used in the LTTng US
trace (stream). To avoid the parsing of the CTF metadata, we add a
sink that extracts the relevant clock definitions from the stream
beginning messages and emits them as JSON lines on a file descriptor.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>