feat(tool): support combining ftrace and LTTng US traces
We first acquire data about the clock of the LTTng US trace, which
contains a reference point to the world clock. Then, we use that
information to create the ftrace clock accordingly and finally mux the
traces.

By that, the aligned traces can directly be viewed in tracecompass
without any further postprocessing.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>