chore: add option to allow usage of libtracecmd private symbols
The trace.dat file contains information that is important to map the
kernel trace to a userland trace. However, the APIs to access this
information is only exposed via exported symbols, but not officially
part of the API.

To deal with this until this has been clarified upstream, we add a build
option to allow the usage of a limited set of private libtracecmd
functions.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>