feat: pass trace metadata from US trace to kernel trace
The sink.ctf.fs sink has logic that detects if a processed trace is an
LTTng trace. If it is, the output directory is layed out as done by
LTTng. To support this, we need to pass metadata from the LTTng US trace
into the ftrace source. This is implemented in a two staged approach:
Metadata we can read from the ftrace itself is just added. Metadata
(like the trace-name) that is only available in the LTTng US trace is
copied over.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>