fix(plugin): correctly free trace record on interruption
Trace records need to be freed with tracecdm_free_record instead of
free. This was done incorrectly when finalizing the message iterator.
However, the bug remained unnoticed as the record is NULL on full
iteration, hence the free() is simply a noop.

On partial iteration, the record was not properly freed, leading to a
leak and debug output.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>