chore(tool): ignore discarded events
Before we can emit discarded events, we need to ensure these are
compatible with the CTF sink. While we can correctly emit these events
as reported by ftrace, the semantics of the reported time-ranges do not
match the CTF (LTTng) expectations, hence cannot be handled by the CTF
sink. As we still want to have these events for other sinks, we
configure the CTF sink to ignore discarded events.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>