chore(plugin): only create ports for streams with data
In case we have event streams of CPUs that did not see any events, these
are empty. These empty streams create downstream problems (e.g. in
trace-compass). As we easily can check if we have events, we do that
upfront and only create a port if we at least have one event in the
flight record.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>